refactor: add publication folder on store
