+++ apps/web/src/components/Composer/Actions/CollectSettings/CollectLimitConfig.tsx
@@ -3,7 +3,7 @@ import type { FC } from 'react';
 import ToggleWithHelper from '@components/Shared/ToggleWithHelper';
 import { StarIcon } from '@heroicons/react/24/outline';
 import { Input } from '@hey/ui';
-import { useCollectModuleStore } from 'src/store/non-persisted/useCollectModuleStore';
+import { useCollectModuleStore } from 'src/store/non-persisted/publication/useCollectModuleStore';
 
 interface CollectLimitConfigProps {
   setCollectType: (data: any) => void;
