+++ apps/web/src/components/Composer/Actions/CollectSettings/FollowersConfig.tsx
@@ -2,7 +2,7 @@ import type { FC } from 'react';
 
 import ToggleWithHelper from '@components/Shared/ToggleWithHelper';
 import { UserGroupIcon } from '@heroicons/react/24/outline';
-import { useCollectModuleStore } from 'src/store/non-persisted/useCollectModuleStore';
+import { useCollectModuleStore } from 'src/store/non-persisted/publication/useCollectModuleStore';
 
 interface FollowersConfigProps {
   setCollectType: (data: any) => void;
