+++ apps/web/src/components/Composer/Actions/CollectSettings/index.tsx
@@ -4,7 +4,7 @@ import { RectangleStackIcon } from '@heroicons/react/24/outline';
 import { Modal, Tooltip } from '@hey/ui';
 import { motion } from 'framer-motion';
 import { useState } from 'react';
-import { useCollectModuleStore } from 'src/store/non-persisted/useCollectModuleStore';
+import { useCollectModuleStore } from 'src/store/non-persisted/publication/useCollectModuleStore';
 
 import CollectForm from './CollectForm';
